+++ src/realm/parser/collection_operator_expression.hpp
@@ -125,8 +125,9 @@ struct CollectionOperatorGetter {
 };
 
 template <typename RetType>
-struct CollectionOperatorGetter<RetType, parser::Expression::KeyPathOp::Min,
-                                typename std::enable_if_t<realm::is_any<RetType, Int, Float, Double>::value>> {
+struct CollectionOperatorGetter<
+    RetType, parser::Expression::KeyPathOp::Min,
+    typename std::enable_if_t<realm::is_any<RetType, Int, Float, Double, Decimal128>::value>> {
     static SubColumnAggregate<RetType, aggregate_operations::Minimum<RetType> > convert(const CollectionOperatorExpression<parser::Expression::KeyPathOp::Min>& expr)
     {
         if (expr.pe.dest_type_is_backlink()) {
@@ -145,8 +146,9 @@ struct CollectionOperatorGetter<RetType, parser::Expression::KeyPathOp::Min,
 };
 
 template <typename RetType>
-struct CollectionOperatorGetter<RetType, parser::Expression::KeyPathOp::Max,
-                                typename std::enable_if_t<realm::is_any<RetType, Int, Float, Double>::value>> {
+struct CollectionOperatorGetter<
+    RetType, parser::Expression::KeyPathOp::Max,
+    typename std::enable_if_t<realm::is_any<RetType, Int, Float, Double, Decimal128>::value>> {
     static SubColumnAggregate<RetType, aggregate_operations::Maximum<RetType> > convert(const CollectionOperatorExpression<parser::Expression::KeyPathOp::Max>& expr)
     {
         if (expr.pe.dest_type_is_backlink()) {
@@ -165,8 +167,9 @@ struct CollectionOperatorGetter<RetType, parser::Expression::KeyPathOp::Max,
 };
 
 template <typename RetType>
-struct CollectionOperatorGetter<RetType, parser::Expression::KeyPathOp::Sum,
-                                typename std::enable_if_t<realm::is_any<RetType, Int, Float, Double>::value>> {
+struct CollectionOperatorGetter<
+    RetType, parser::Expression::KeyPathOp::Sum,
+    typename std::enable_if_t<realm::is_any<RetType, Int, Float, Double, Decimal128>::value>> {
     static SubColumnAggregate<RetType, aggregate_operations::Sum<RetType> > convert(const CollectionOperatorExpression<parser::Expression::KeyPathOp::Sum>& expr)
     {
         if (expr.pe.dest_type_is_backlink()) {
@@ -185,8 +188,9 @@ struct CollectionOperatorGetter<RetType, parser::Expression::KeyPathOp::Sum,
 };
 
 template <typename RetType>
-struct CollectionOperatorGetter<RetType, parser::Expression::KeyPathOp::Avg,
-                                typename std::enable_if_t<realm::is_any<RetType, Int, Float, Double>::value>> {
+struct CollectionOperatorGetter<
+    RetType, parser::Expression::KeyPathOp::Avg,
+    typename std::enable_if_t<realm::is_any<RetType, Int, Float, Double, Decimal128>::value>> {
     static SubColumnAggregate<RetType, aggregate_operations::Average<RetType> > convert(const CollectionOperatorExpression<parser::Expression::KeyPathOp::Avg>& expr)
     {
         if (expr.pe.dest_type_is_backlink()) {
@@ -205,8 +209,9 @@ struct CollectionOperatorGetter<RetType, parser::Expression::KeyPathOp::Avg,
 };
 
 template <typename RetType>
-struct CollectionOperatorGetter<RetType, parser::Expression::KeyPathOp::Count,
-                                typename std::enable_if_t<realm::is_any<RetType, Int, Float, Double>::value>> {
+struct CollectionOperatorGetter<
+    RetType, parser::Expression::KeyPathOp::Count,
+    typename std::enable_if_t<realm::is_any<RetType, Int, Float, Double, Decimal128>::value>> {
     static LinkCount convert(const CollectionOperatorExpression<parser::Expression::KeyPathOp::Count>& expr)
     {
         if (expr.pe.dest_type_is_backlink()) {
@@ -222,8 +227,9 @@ struct CollectionOperatorGetter<RetType, parser::Expression::KeyPathOp::Count,
 
 
 template <typename RetType>
-struct CollectionOperatorGetter<RetType, parser::Expression::KeyPathOp::BacklinkCount,
-                                typename std::enable_if_t<realm::is_any<RetType, Int, Float, Double>::value>> {
+struct CollectionOperatorGetter<
+    RetType, parser::Expression::KeyPathOp::BacklinkCount,
+    typename std::enable_if_t<realm::is_any<RetType, Int, Float, Double, Decimal128>::value>> {
     static BacklinkCount<Int>
     convert(const CollectionOperatorExpression<parser::Expression::KeyPathOp::BacklinkCount>& expr)
     {