+++ CHANGELOG.md
@@ -1,11 +1,11 @@
 # NEXT RELEASE
 
 ### Enhancements
-* None.
+* Added query parser support for Decimal128 and ObjectId types.
 
 ### Fixed
-* <How to hit and notice issue? what was the impact?> ([#????](https://github.com/realm/realm-core/issues/????), since v?.?.?)
-* None.
+* Fixed float and double maximum queries when all values were less or equal to zero.
+* Fixes parsing float and double constants which had been serialised to scientific notation (eg. 1.23E-24). ([#3076](https://github.com/realm/realm-core/issues/3076)).
  
 ### Breaking changes
 * None.